Venear governance tests
